refactor: split runner.py into runner_models.py and repo_ops.py
Extract data models (InstallerBackend, RunnerConfig, PackageResult,
RunSummary, RunOutput) to runner_models.py and git operations, package
spec parsing, and sdist helpers to repo_ops.py. All names are re-exported
from runner.py using the `as` idiom so existing imports and mock.patch
targets continue to work unchanged.

runner.py reduced from 1972 to 1445 lines (~27%).

Closes #136
